Backend/requirements.txt
flask
flask-cors
psycopg2
Flask-Migrate
python-dotenv